requirements.txt
# requirements.txt

PyYAML==6.0.1
